requirements.txt
black==23.3.0
Pillow==9.5.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
